Added log-in data to pg.js, moved controllers & models files
