+++ server/models/main_functions/queries.js
@@ -43,3 +43,5 @@ where user_email = $1;
 `
 
 }
+
+//just trying to get git to recognize this file
